backend/requirements.txt
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.2.0
pillow>=10.2.0
aiosqlite>=0.19.0
asyncpg>=0.30.0
boto3>=1.35.0
celery[redis]>=5.4.0
redis>=5.2.0
